requirements.txt
pyftpdlib
libtero
django
channels
asgi_redis
